+++ test/_helpers.ts
@@ -1,6 +1,4 @@
-// TODO: Fix ts-ignore
-// @ts-ignore
-import dataUriToBuffer from 'data-uri-to-buffer'
+const dataUriToBuffer = require('data-uri-to-buffer')
 import dataUriToBlob from 'dataurl-to-blob'
 
 export const dataURItoBuffer: (uri: string) => Buffer = dataUriToBuffer
@@ -15,6 +13,4 @@ export const isNode = (): boolean => {
   }
 }
 
-// TODO: Fix ts-ignore
-// @ts-ignore
 export const sleep = (ms: number): Promise<void> => new Promise(resolve => setTimeout(resolve, ms))